Adding support for macOS
This changes a lot of stuff to support macOS. Architecture stays the same. There are a lot of platform dependent code paths to support iOS and OSX.

Most of the things were done with `#if os(some OS here)`.

The podspec adds the correct files for each platform.
The Xcode project has 2 seperate frameworks, one for iOS and one for macOS.

Demo projects also were updated to reflect the api and the available content views, equivalent on both platforms (though on iOS not all the content views are shown, on macOS all are demoed)

IMPORTANT NOTE:
On PKHUD init the view on which the HUD should be presented on can now be set. Otherwise the view can be set afterwards through `viewToPresentOn`.
If none is set the implementation looks for the current window or any window to present the hud on. On macOS all `orderedWindows.first?.contentView` is taken as backup view. On iOS the framework searches for the `keyWindow` or `windows.first` on `sharedApplication`.

On macOS in particular it is important to set `viewToPresentOn` when the HUD will be presented some time later (i.e. async as part of a callback or on a signal), because in the time between the user could switch windows and the HUD then would be presented on the current window instead of on the intended window.